+++ Package.swift
@@ -1,0 +1,23 @@
+//
+// ValueCoding	
+// File created on 14/09/2017.
+// 	
+// Copyright (c) 2015-2017 Daniel Thorpe
+// 
+// ValueCoding is licensed under the MIT License. Read the full license at 
+// https://github.com/danthorpe/ValueCoding/blob/master/LICENSE
+//
+
+// swift-tools-version:4.0
+
+import PackageDescription
+
+let package = Package(
+    name: "ValueCoding",
+    targets: [
+        Target(name: "ValueCoding")
+    ],
+    exclude: [
+        "Tests"
+    ]
+)